+++ components/camel-jms/src/test/java/org/apache/camel/component/jms/AbstractPersistentJMSTest.java
@@ -48,12 +48,4 @@ public abstract class AbstractPersistentJMSTest extends CamelTestSupport {
         ConnectionFactory connectionFactory = ConnectionFactoryHelper.createConnectionFactory(service);
         camelContext.addComponent("activemq", jmsComponentAutoAcknowledge(connectionFactory));
     }
-
-    protected void waitForJmsConsumerRoutes(String... routeIds) {
-        AbstractJMSTest.waitForJmsConsumerRoutes(context, routeIds);
-    }
-
-    protected void waitForJmsConsumerRoutes(long minUptimeMillis, String... routeIds) {
-        AbstractJMSTest.waitForJmsConsumerRoutes(context, minUptimeMillis, routeIds);
-    }
 }
